Move cloneurls package to frontend
Only used here, and it should not be used in gitserver so this is an easy way to ensure the right cloneurl package is used.

Test plan:

Just moved code around, CI will find issues.
