+++ cmd/frontend/graphqlbackend/BUILD.bazel
@@ -238,6 +238,7 @@ go_library(
         "//cmd/frontend/graphqlbackend/graphqlutil",
         "//cmd/frontend/hubspot",
         "//cmd/frontend/hubspot/hubspotutil",
+        "//cmd/frontend/internal/cloneurls",
         "//cmd/frontend/internal/processrestart",
         "//cmd/frontend/internal/search/logs",
         "//internal/actor",
@@ -249,7 +250,6 @@ go_library(
         "//internal/authz",
         "//internal/authz/permssync",
         "//internal/binary",
-        "//internal/cloneurls",
         "//internal/cloud",
         "//internal/codeintel/dependencies",
         "//internal/codeintel/dependencies/shared",
