+++ cmd/frontend/internal/app/BUILD.bazel
@@ -30,6 +30,7 @@ go_library(
         "//cmd/frontend/internal/app/router",
         "//cmd/frontend/internal/app/ui",
         "//cmd/frontend/internal/auth/accessrequest",
+        "//cmd/frontend/internal/cloneurls",
         "//cmd/frontend/internal/routevar",
         "//cmd/frontend/oneclickexport",
         "//internal/actor",
@@ -37,7 +38,6 @@ go_library(
         "//internal/auth",
         "//internal/auth/userpasswd",
         "//internal/authz",
-        "//internal/cloneurls",
         "//internal/conf",
         "//internal/conf/deploy",
         "//internal/cookie",
