+++ cmd/frontend/internal/repos/webhooks/BUILD.bazel
@@ -8,9 +8,9 @@ go_library(
     visibility = ["//cmd/frontend:__subpackages__"],
     deps = [
         "//cmd/frontend/enterprise",
+        "//cmd/frontend/internal/cloneurls",
         "//cmd/frontend/internal/repos/webhooks/resolvers",
         "//cmd/frontend/webhooks",
-        "//internal/cloneurls",
         "//internal/codeintel",
         "//internal/conf/conftypes",
         "//internal/database",
